Refactor, split problems and optmizers, add problem_solver
